Move the account migration screen to within the room list
