+++ ElementX/Sources/UITests/UITestsScreenIdentifier.swift
@@ -27,7 +27,6 @@ enum UITestsScreenIdentifier: String {
     case waitlist
     case analyticsPrompt
     case analyticsSettingsScreen
-    case migration
     case templateScreen
     case appLockFlow
     case appLockFlowAlternateWindow
